Refactors into Send
