+++ src/view/UnlockWallet.svelte
@@ -7,6 +7,8 @@
     import LabelledLoader from "../components/LabelledLoader.svelte";
     import {onMount} from "svelte";
     import {softwareKeysStore} from "../stores/stores";
+    import {clearSoftwareKeys, setSoftwareKeys} from "../machinery/SoftwareKeysState";
+    import type {SoftwareKeysState} from "../machinery/SoftwareKeysState";
 
     let inputPhrase: string | undefined;
     let showLoader: boolean = false;
@@ -16,6 +18,7 @@
     }
 
     const unlock = async () => {
+        clearSoftwareKeys();
         showLoader = true;
         const data: NanoWallet | undefined = await unlockWallet(inputPhrase)
         if(data) {
@@ -23,23 +26,24 @@
             pushState({menu: 'wallet', wallet: data, account: undefined, onboardState: undefined})
         } else {
             pushToast({ languageId: 'wrong-pass' })
+            setSoftwareKeys(softwareKeys)
         }
         showLoader = false;
     }
 
-    onMount(() => {
-        softwareKeysStore.set({
-            leftKey: {
-                onClick: () => pushMenu('onboard'),
-                languageId: 'create-new-wallet'
-            },
-            middleKey: {
-                onClick: unlock,
-                languageId: 'unlock-wallet'
-            },
-            rightKey: undefined
-        })
-    })
+    const softwareKeys: SoftwareKeysState = {
+        leftKey: {
+            onClick: () => pushMenu('onboard'),
+            languageId: 'create-new-wallet'
+        },
+        middleKey: {
+            onClick: unlock,
+            languageId: 'unlock-wallet'
+        },
+        rightKey: undefined
+    }
+
+    onMount(() => setSoftwareKeys(softwareKeys))
 
 </script>
